lib/storage: typo fixes found by golangci-lint; updates https://github.com/VictoriaMetrics/VictoriaMetrics/issues/69

+++ lib/storage/index_db.go
@@ -250,7 +250,7 @@ func (db *indexDB) incRef() {
 
 func (db *indexDB) decRef() {
 	n := atomic.AddUint64(&db.refCount, ^uint64(0))
-	if n < 0 {
+	if int64(n) < 0 {
 		logger.Panicf("BUG: negative refCount: %d", n)
 	}
 	if n > 0 {
